task 1/3: Extract preview mocks and consolidate the 5 @Preview functions into preview/SwapPreviewData.kt (#4836)
